rename from-typescript-class.ts
