+++ lib/typescript-type-introspector/tree-sitter-typescript-type-introspector.ts
@@ -1,6 +1,6 @@
 import type { TSchema } from "@sinclair/typebox";
 import type { Parser } from "#/lib/tree-sitter/tree-sitter.ts";
-import { getTypeBoxSchemaFromTreeSitterTypeScriptClass } from "#/lib/tree-sitter/tree-sitter-typebox/typescript-class.ts";
+import { getTypeBoxSchemaFromTreeSitterTypeScriptClass } from "#/lib/tree-sitter/tree-sitter-typebox/from-typescript-class.ts";
 import type { TypeScriptTypeIntrospector } from "./typescript-type-introspector.ts";
 
 /**
